pyproject.toml
[tool.poetry]
name = "fitnest-django-backend"
version = "0.1.0"
description = ""
authors = ["falakhe <fsivela@fragomen.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
drf-writable-nested = "^0.7.0"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular-sidecar = "^2024.7.1"
django-storages = "^1.14.4"
boto3 = "^1.35.14"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
